Created base class for knowledge distillation
